+++ mobile/src/tabs/index.js
@@ -1,5 +1,5 @@
 export { default as ConnectScreen, ConnectStack } from './connect';
+export { default as HomeScreen, HomeStack } from './home';
 export { default as ProfileScreen, ProfileStack } from './profile';
 export { default as SearchScreen, SearchStack } from './search';
 export { default as SectionsScreen, SectionsStack } from './sections';
-// export { default as Feed } from './Feed';
